+++ packages/highlight/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @lmarcel/highlight
 
+## 2.1.1
+
+### Patch Changes
+
+- Fixing bug on call edit mode functions and improving storybook documentation
+
 ## 2.1.0
 
 ### Minor Changes